v1b2 - false negative: 403 errors weren't being reported or changing behaviour for second download attempts, after first 403 errors made attempt with slightly modified URL. That's been fixed!

+++ dart
@@ -2,7 +2,7 @@
 # encoding: utf-8
 
 # Name: iStore Artwork Downloader - dArt - download Art
-# Version: 1b1
+# Version: 1b2
 # Purpose: This application will search Apple's online stores (iTunes Store, iBookstore, Mac App Store and iOS App Store) for a user-input term and retrieve the full resolution cover art images from Apple's servers. It MAY support other sources in the future, but for now, it's limited to Apple's online media stores.
 # Author: Habib alAmin
 
@@ -184,12 +184,18 @@ try:
                 urlretrieve(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100-75", ""), expandvars(expanduser(join(downloadLocation, replace(searchResults[u'results'][i][u'collectionName'], "/", "--")))) + splitext(searchResults[u'results'][i][u'artworkUrl100'])[1])
                 print "The artwork for \"" + searchResults[u'results'][i][u'collectionName'] + "\"", "has been retrieved!"
         elif urlopen(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100-75", "")).getcode() == 403:
-            try:
-                urlretrieve(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100", ".600x600"), expandvars(expanduser(join(downloadLocation, replace(searchResults[u'results'][i][u'trackName'], "/", "--")))) + splitext(searchResults[u'results'][i][u'artworkUrl100'])[1])
-                print "The artwork for \"" + searchResults[u'results'][i][u'trackName'] + "\"", "has been retrieved!"
-            except KeyError:
-                urlretrieve(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100", ".600x600"), expandvars(expanduser(join(downloadLocation, replace(searchResults[u'results'][i][u'collectionName'], "/", "--")))) + splitext(searchResults[u'results'][i][u'artworkUrl100'])[1])
-                print "The artwork for \"" + searchResults[u'results'][i][u'collectionName'] + "\"", "has been retrieved!"
+            if urlopen(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100-75", ".600x600")).getcode() == 200:
+                try:
+                    urlretrieve(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100", ".600x600"), expandvars(expanduser(join(downloadLocation, replace(searchResults[u'results'][i][u'trackName'], "/", "--")))) + splitext(searchResults[u'results'][i][u'artworkUrl100'])[1])
+                    print "The artwork for \"" + searchResults[u'results'][i][u'trackName'] + "\"", "has been retrieved!"
+                except KeyError:
+                    urlretrieve(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100", ".600x600"), expandvars(expanduser(join(downloadLocation, replace(searchResults[u'results'][i][u'collectionName'], "/", "--")))) + splitext(searchResults[u'results'][i][u'artworkUrl100'])[1])
+                    print "The artwork for \"" + searchResults[u'results'][i][u'collectionName'] + "\"", "has been retrieved!"
+            else:
+                try:
+                    print "\033[91mThe artwork for \"" + searchResults[u'results'][i][u'trackName'] + "\"", "could not be retrieved!" + "\033[1m", urlopen(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100-75", ".600x600")).getcode(), "error!\033[0m"
+                except KeyError:
+                    print "\033[91mThe artwork for \"" + searchResults[u'results'][i][u'collectionName'] + "\"", "could not be retrieved!" + "\033[1m", urlopen(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100-75", ".600x600")).getcode(), "error!\033[0m"
         else:
             try:
                 print "\033[91mThe artwork for \"" + searchResults[u'results'][i][u'trackName'] + "\"", "could not be retrieved!" + "\033[1m", urlopen(replace(searchResults[u'results'][i][u'artworkUrl100'], ".100x100-75", "")).getcode(), "error!\033[0m"
